serve: extract config module
Move the env-var parsing and *_from_env / parse_* helpers, plus the
context-size (initial/OOM-shrink) arithmetic, out of src/cmd/serve.rs
into src/cmd/serve/config.rs, following the existing aggregation.rs /
responses.rs submodule pattern.

Pure move-code refactor — no behaviour changes. Part of the tracking
issue to split the oversized serve.rs (llmmanorg/llmman#429).
